Move test to test_actor.py
