+++ python/ray/tests/utils.py
@@ -81,3 +81,16 @@ def run_string_as_driver_nonblocking(driver_script):
         f.flush()
         return subprocess.Popen(
             [sys.executable, f.name], stdout=subprocess.PIPE)
+
+
+def relevant_errors(error_type):
+    return [info for info in ray.error_info() if info["type"] == error_type]
+
+
+def wait_for_errors(error_type, num_errors, timeout=10):
+    start_time = time.time()
+    while time.time() - start_time < timeout:
+        if len(relevant_errors(error_type)) >= num_errors:
+            return
+        time.sleep(0.1)
+    raise Exception("Timing out of wait.")
